updated 1.2 - TWP130 elif.py
I just changed 'float' in the 7 line for 'int'.The program worked fine.

+++ Python para Zumbis/2 - Mastigando estruturas de controle/Criando condicoes em seu programa/1.1 - TWP130 elif.py
@@ -4,7 +4,7 @@ Entre 200 e 400 minutos, o preço é de R$0,18. Acima de 400 minutos o preço po
 Calcule sua conta de telefone.
 '''
 
-minutos = float(input('Quantidade de minutos: '))
+minutos = int(input('Quantidade de minutos: '))
 if minutos < 200:
     preco = 0.20
 elif minutos <= 400:
